web/agent/pyproject.toml
[project]
name = "vpi-1000"
version = "1.0.0"
description = "VPI-1000"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pydantic-ai>=0.1.0",
    "fastapi>=0.115.5,<1.0.0",
    "uvicorn>=0.29.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]
